Refatorar app em blueprints e ajustar testes
