+++ setup.sh
@@ -14,7 +14,6 @@ pip install -r requirements.txt
 flask db upgrade
 
 # Seed initial data (optional but recommended)
-python seed_users.py
-python seed_organizacao.py
+python seed.py
 
 echo "Setup complete."
