+++ tests/test_article_permissions.py
@@ -1,18 +1,15 @@
-import os
 import pytest
 
-os.environ['SECRET_KEY'] = 'test_secret'
-os.environ['DATABASE_URI'] = 'sqlite:///:memory:'
 
 from app import app, db
 from models import Instituicao, Estabelecimento, Setor, Celula, Funcao, User, Article, ArticleStatus
 from enums import Permissao
 
 @pytest.fixture
-def client():
-    app.config['TESTING'] = True
+def client(app_ctx):
+    
     with app.app_context():
-        db.create_all()
+        
         inst = Instituicao(nome='Inst')
         db.session.add(inst)
         db.session.flush()
@@ -26,11 +23,11 @@ def client():
         db.session.add(cel)
         db.session.commit()
         ids = {'est': est.id, 'setor': setor.id, 'cel': cel.id}
-        with app.test_client() as client:
+        with app_ctx.test_client() as client:
             client.base_ids = ids
             yield client
-        db.session.remove()
-        db.drop_all()
+        
+        
 
 def login_user(client, perms=None):
     perms = perms or []
