+++ tests/test_celula.py
@@ -1,17 +1,14 @@
-import os
 import pytest
 
-os.environ.setdefault('SECRET_KEY', 'test_secret')
-os.environ.setdefault('DATABASE_URI', 'sqlite:///:memory:')
 
 from app import app, db
 from models import Instituicao, Estabelecimento, Setor, Celula, User, Funcao
 
 @pytest.fixture
-def client():
-    app.config['TESTING'] = True
+def client(app_ctx):
+    
     with app.app_context():
-        db.create_all()
+        
         inst = Instituicao(nome='Inst')
         db.session.add(inst)
         db.session.flush()
@@ -21,10 +18,10 @@ def client():
         setor = Setor(nome='Setor 1', estabelecimento_id=est.id)
         db.session.add(setor)
         db.session.commit()
-        with app.test_client() as client:
+        with app_ctx.test_client() as client:
             yield client
-        db.session.remove()
-        db.drop_all()
+        
+        
 
 def login_admin(client):
     with app.app_context():
