+++ tests/test_estabelecimento.py
@@ -1,25 +1,22 @@
-import os
 import pytest
 
 # Configure environment variables before importing the app
-os.environ.setdefault('SECRET_KEY', 'test_secret')
-os.environ.setdefault('DATABASE_URI', 'sqlite:///:memory:')
 
 from app import app, db
 from models import Estabelecimento, Instituicao, Setor, Celula, User, Funcao
 
 @pytest.fixture
-def client():
-    app.config['TESTING'] = True
+def client(app_ctx):
+    
     with app.app_context():
-        db.create_all()
+        
         inst = Instituicao(nome='Inst 1')
         db.session.add(inst)
         db.session.commit()
-        with app.test_client() as client:
+        with app_ctx.test_client() as client:
             yield client
-        db.session.remove()
-        db.drop_all()
+        
+        
 
 def login_admin(client):
     with app.app_context():
